requirements.txt
colorlog >= 2.6.1, < 3.0.0
google-apitools
gcloud >= 0.15.0
github3.py >= 0.9.6
grpcio-tools >= 1.0.2
kazoo >= 2.2.1
oslo.utils >= 3.4.0
pbr==2.0.0
requests >= 2.10.0, < 3.0.0
ruamel.yaml >= 0.14.5
six >= 1.10.0
taskflow == 1.25.0
yapf >= 0.6.2
